+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Add `Iterable.toMap`, `Iterable.atRandom`, `Iterable.min`, `Iterable.max` and `Iterable.percentile`.
 * Optimize `Range` object creation and error reporting, and add `Iterable.indices`.
 * Add converters `toBiMap`, `toListMultimap`, and `toSetMultimap`.
+* Add `String.removePrefix` and `String.removeSuffix`.
 
 ## 2.8.0